Remove strokes as they didn't work out
The problem is that this isn't bleed on a printed tile
(where I could rely on the edges butting together neatly on
the exact edge lines after cutting). These tiles are actually
being drawn on top of each other, so the bleed edges of the
later-drawn tiles actually go over the earlier-drawn ones and
cause worse gaps.